Added real city names to thedata generation script
