+++ server/index.js
@@ -7,6 +7,7 @@ const bodyParser = require("body-parser");
 
 
 
+
 app.use(bodyParser.json())
 app.use(cors());
 app.use('/', express.static(path.join(__dirname, '../client/dist')))
